fix: Detect top-level project directory name
The module 'genesis.py' assumed the top-level project directory would be
named 'cluster-genesis'. Git allows a project to use any name for the
top-level directory and so should we. After renaming the project GitHub
repository this is especially problematic since the default directory
name is now 'power-up'. This commit changes the 'GEN_PATH' algorithm to
rely on the relative location of the 'genesis.py' module. If
'genesis.py' were to be moved this code would also need to be updated.