fix: Create empty log file to ensure user is owner
If 'teardown' is run before a log file ('logs/gen') is created then the
log file will be created with 'root' user as the owner and without write
privileges for other users. Subsequent attempts to write to the log file
as a non-root user will fail. Tracking a blank log file in the project
repository will ensure it will be created with the cloning user as
owner.