fix: Improve 'pup' output when creating container
When running commands inside the container 'stderr' is now written to
file (along with 'stdout') instead of printing to the user's 'stdout'.
There's no need to clutter the output if return codes from the commands
are good and then script continues running.

Output from container commands are written to a '.log' file inside the
'logs/' directory.