fix: Update service init.d properties
The first commit in this branch was derived through quick trial and
error. I spent a little more time to understand the start/stop
configuration in init.d files. I believe this commit is the correct way
to configure services to automatically start (and stop) when the
container enters certain runlevels.