fix: Add ipmi_set_bootdev back to gen.py
Added ipmi_set_bootdev back into gen.py during inv_add_ports_pxe. The
boot device is set to 'setup' (bios) at this stage. This will
hopefully prevent misses due to existing OS which does not have the
pxe port set to DHCP and nodes which do not obtain an ip during pxe
boot.

Refactored ipmi_set_bootdev to use ipmi addresses if no client list is
supplied. If a client list is supplied it is assumed to be pxe port IP
addresses which are then mapped to IPMI addresses.