feat: Facilitate single PowerAI license acceptance
Users are required to interactively accept the PowerAI license agreement
on at least one server per cluster. A single server is selected
(no significance as to _which_ server) to run the manual acceptance
script. If the user accepts the license agreement on that server then
the acceptance script will be run quietly on the remaining servers.