feat: Prompt for SSH credentials to copy SSH keys
Users are now only required to provide a list of hosts in the Ansible
inventory file ('software_hosts'). If no SSH credentials are provided
then the user will be prompted to 1) create or select an SSH key pair
and 2) provide a username/password (via cli prompts) in order to copy
the SSH public key to all hosts. The Ansible inventory is updated with
the appropriate vars to continue using the selected SSH key. The
interactive configuration only works with a single global SSH user name
and password, but a manually configuring host specific variables in the
Ansible inventory should still work.

The inventory file ('software_hosts') is moved to the
'power-up/software' directory, and it is owned by the "login" user
instead of root.

SSH 'StrictHostKeyChecking' is forced to be disabled.