fix: Always unlock vault before running plays
Any 'pup' process instantiating a 'software' object will delete the
'.vault' file upon termination. If more than one 'pup' process is run
concurrently each process needs to verify the existence of the '.vault'
file before attempting to use it. The vault password is stored as an
attribute and thus can be used to re-unlock without re-prompting the
user.

There is still a small chance that the '.vault' file is removed
in-between checking for it's existence and starting up the
'ansible-playbook' sub-process. In this case a retry will be required.