fix: Install 'ipaddress' python module for Cobbler
The Cobbler web GUI fails to load because of an 'ImportError' for the
module 'ipaddress'. Cobbler uses the system level python installation so
this package needs to be installed outside of the pup python venv. The
standard Ubuntu 14.04 repos do not have a 'python-ipaddress' package
(releases >14.04 do) so the module is installed running 'pip' as root
user.