refactor: Update Ubuntu installer image URLs
URLs for Ubuntu 14.04 for ppc64el is *removed*. The required "mini" iso
containing the installer initfs and kernel is no longer hosted by
ubuntu.com.

Ubuntu 16.04 URLs (and checksums) is updated to point to the 16.04.5
image. It is not recommended to use older (16.04.3, 16.04.4, etc.)
install images so we should not facilitate automatically downloading
them.