feat: Allow post-deploy '--gather-mac-addr' retry
If post-deploy '--gather-mac-addr' (gathers client node data interface
MAC addresses) fails to collect all addresses the user needs to both be
informed and given a recovery path.

If _any_ data interface MAC address is missing a table of _all_ data
interfaces is printed with the following columns: 'Switch', 'Port',
'Device', 'MAC'.

Once the table is printed the user is prompted to retry, continue (with
the missing MAC addresses), or exit the program.