refactor: Display install script messages at end
The install script, 'install.sh', prints lots of lines to stdout making
it difficult for a user to notice any warning or error messages. Using a
global 'MESSAGES' variable allows stashing messages and printing them
all at the end. Message are still printed to stdout as they're created
so they not get lost in the case of the script exiting due to an error.