fix: Disable stale PXE ifcfg-* script if renaming
Since the PXE interface is using during OS installation a 'ifcfg-*'
script is automatically generated. If this interface is renamed and
reconfigured during post-deploy the original 'ifcfg-*' script is left in
place and causes systemd to report errors when starting
'network.service'. Instead of removing the 'ifcfg-*' script this patch
sets 'ONBOOT=no' so that 'network.service' does not try to activate it.

I have not tested Ubuntu to see if similar behavior exists there. This
patch only affects RHEL client nodes.