fix: Run data switch config script on host
There is a remnant 'lxc-ls' command in a 'genesis' function to check if
the container is running. This function was only used in some logic to
run the data switch configuration script inside or outside of the
container. The script can safely be run outside the container (directly
on the deployer host) all of the time so the logic should be removed
entirely.

There is also a 'genesis' function to get the container IP address
('container_addr()') that uses 'lxc-ls'. This function is only used the
in the 'show_status.py' script, which needs addtional cleanup. I'll
leave the 'show_status.py' cleanup to a future commit.