fix: Exit yum with error if package is missing
When installing multiple packages at the same time yum will still exit
cleanly if some of the packages are missing. There is an option to force
an exit with error if _any_ package is missing. All packages installed
via the 'scripts/install.sh' script are required and thus the software
should halt if it is unable to install something.