fix: Remove module level 'logger.create()'
A 'logger.create()' inside of the 'osinstall' module resulted in
multiple logging objects (and multiple messages printed to stdout).

The logger class is singleton and should only be created once: using
argparse inside of a '__main__' function. Elsewhere the 'getlogger()'
function can be used to get a reference to the existing logger object.