fix: Decrement BMC credential count
When validating BMC IPMI/OpenBMC credentials a count is stored to keep
track of the successful connections. When a new successful connection is
validated the count is decremented by the _total_ number of BMCs instead
of 1. This resulted in a binary pass/fail if _any_ or _no_ BMC
connections were validated.