refactor: Enhance yes/no prompts
Enhance get_yesno to allow a default empty response which is translated
to a yes or no based on the default yesno arg.

Enhance wmla120.py to use a default empty no prompt 'y/[n]' if a
given content or repo is already set up, and to use a default empty
yes prompt '[y]/n' if the content or repo is not setup.