feat: Allow 'wmla120' self-installation
Allow including installer host in the inventory for a "self-install".
The eliminated the need for a dedicated installer node.

Rebooting is controlled via an Ansible variable, 'pup_reboot', that is
set in the inventory. A global 'pup_reboot=True' is automatically added
to default to original reboot behavior. If the installer node is
included in the inventory, a 'pup_reboot=True' host variable is
automatically added to the inventory (and anytime validation is called
it will ensure this value is set, preventing an override). Additional
client nodes could also set 'pup_reboot=True' to prevent them from
rebooting.