fix: Install WMLA license to root home dir
The WMLA license acceptance script ('accept-ibm-wmla-license.sh') is run
though an SSH subprocess of the main 'wmla120.py' module. Directory
paths need to be consistent between the acceptance script and the
verification run through Ansible. The 'HOME' environmental variable  is
used in this commit to achieve that consistency.

Since the acceptance file ('status.dat') is located in a directory owned
by root, privilege escalation is required just to check if the license
is already installed. There seems to be no harm in re-running the
acceptance script, so the simplest approach is to prompt for the 'sudo'
password once and run the acceptance script without checking to see if
it's already been accepted.