fix: Use 'replace_regex' to set 'pup_reboot' var
The 'pup_reboot' Ansible variable is used to control which client nodes
are automatically rebooted (e.g. can prevent the installer node from
self reboot in the middle of running pup). The logic to set 'pup_reboot'
failed to account for whitespace between beginning of line and hostname,
user defined variables and comments.