fix: Update pypi repo build error mask
When downloading python packages from a pypi repository it is only
necessary to get the '*.tar.gz' archive. If there is a build error
(could be due to missing system dependencies on the installer, incorrect
python environment, etc.) it is safe to ignore since this it is not in
an equivalent runtime environment.